add styles

+++ src/components/nweet.js
@@ -1,6 +1,11 @@
 import { dbService, storageService } from "fbase";
 import React, { useState } from "react";
 import { doc, updateDoc, deleteDoc } from "firebase/firestore";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import stylesNweet from "../style/nwitter.module.css";
+import { faTrash } from "@fortawesome/free-solid-svg-icons";
+import { faPenToSquare } from "@fortawesome/free-regular-svg-icons";
+import stylesedit from "../style/edit.module.css";
 
 const Nweet = ({ nweetObj, isOwner }) => {
   const [editing, setEditing] = useState(false);
@@ -12,7 +17,8 @@ const Nweet = ({ nweetObj, isOwner }) => {
     const ok = window.confirm("Are you sure you want to delete this nweet?");
     if (ok) {
       await deleteDoc(nweetTextRef);
-      await storageService.refFromURL(nweetObj.attachmentUrl).delete();
+      if (nweetObj.attachmentUrl)
+        await storageService.refFromURL(nweetObj.attachmentUrl).delete();
     } else {
     }
   };
@@ -29,34 +35,50 @@ const Nweet = ({ nweetObj, isOwner }) => {
     setEditing(false);
   };
   return (
-    <div>
+    <div className={stylesNweet.nweetBox}>
       {editing ? (
-        <>
+        <div className={stylesedit.editBox}>
           <form onSubmit={onSubmit}>
             <input
+              className={stylesedit.modifyNweet}
               onChange={onChange}
               type="text"
               placeholder="Edit your nweet"
               value={newNweet}
+              maxLength={120}
               required
             />
-            <input type="submit" value="Modify" />
+            <input
+              type="submit"
+              value="Modify"
+              className={stylesedit.modifySub}
+            />
           </form>
-          <button onClick={toggleEditing}>Cancel</button>
-        </>
+          <button onClick={toggleEditing} className={stylesedit.cancelBtn}>
+            Cancel
+          </button>
+        </div>
       ) : (
-        <>
+        <div className={stylesNweet.nweetText}>
           <h4>{nweetObj.text}</h4>
-          {nweetObj.attachmentUrl && (
-            <img src={nweetObj.attachmentUrl} width="50px" height="50px" />
-          )}
+          {nweetObj.attachmentUrl && <img src={nweetObj.attachmentUrl} />}
           {isOwner && (
-            <>
-              <button onClick={onDelClick}>Delete Nweet</button>
-              <button onClick={toggleEditing}>Edit Nweet</button>
-            </>
+            <div>
+              <button onClick={onDelClick}>
+                <FontAwesomeIcon
+                  icon={faTrash}
+                  className={stylesNweet.deleteLogo}
+                />
+              </button>
+              <button onClick={toggleEditing}>
+                <FontAwesomeIcon
+                  icon={faPenToSquare}
+                  className={stylesNweet.editLogo}
+                />
+              </button>
+            </div>
           )}
-        </>
+        </div>
       )}
     </div>
   );
